refactor: MOVED method generate_password() to its own module in ftapi/utils, and added test.
