+++ ftapi/users/create_users.py
@@ -4,8 +4,6 @@
 
 import json
 import re
-import secrets
-import string
 import sys
 from time import sleep
 from FtApi import FtApi
@@ -14,27 +12,7 @@ from FtCursus import compute_cursus_end_from_begin
 from FtDateTime import is_valid_date
 from FtUser import FtUser, FtCursusUser
 from FtUtils import ft_write_error, ft_write_info, ft_write_success
-
-
-def generate_password(length: int = 12) -> str:
-    """Generates a random password of a given length
-
-    Attributes:
-        length (int): Length of the password to generate. Default is 12."""
-    allowed_chars = [string.ascii_uppercase, "!@#$%^*()_+"]
-    password = [secrets.choice(allowed_chars[0]) for n in range(length//4)]
-    password += [secrets.choice(allowed_chars[1]) for n in range(length//4)]
-    length = length - 2 * (length // 4)
-    allowed_chars = [string.ascii_lowercase,
-                     string.digits,
-                     "-=;,./"]
-    password += [secrets.choice(allowed_chars[0]) for n in range(length//2)]
-    length = length - (length // 2)
-    password += [secrets.choice(allowed_chars[1]) for n in range(length//2)]
-    length = length - (length // 2)
-    password += [secrets.choice(allowed_chars[2]) for n in range(length)]
-    password = ''.join(password)
-    return password
+from utils.generate_password import generate_password
 
 
 def create_user(ft_user: FtUser = None, ft_api: FtApi = None):
